requirements.txt
click==8.1.7
numpy==1.26.1
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
PySide6-Essentials==6.6.0
python-dotenv==1.0.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
shiboken6==6.6.0
